requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dateutil>=2.8.2

# NLP and semantic search
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
transformers>=4.30.0
torch>=2.0.0

# Web interface
streamlit>=1.28.0
gradio>=3.40.0
flask>=2.3.0

# Data visualization
plotly>=5.15.0
altair>=5.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Database (optional)
sqlite3
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0
nltk>=3.8.0

# Development and testing
pytest>=7.4.0
jupyter>=1.0.0
ipykernel>=6.25.0